requirements.txt
fastapi
uvicorn[standard]
starlette
anyio
requests
httpx
apscheduler
python-dotenv
psutil
pandas
numpy
python-dateutil
shapely
beautifulsoup4
feedparser
lxml
loguru
typing-extensions